refactor: split tool-registry into domain sub-registries + block sisyphus-junior
W3 from architectural plan: extract createToolRegistry (241 lines, 8 params)
into 5 focused sub-registries:
- search-tools.ts: grep, glob, ast_grep, session_manager, builtins
- background-tools.ts: background tasks, interactive_bash
- delegation-tools.ts: task, call_omo_agent, look_at, a2a_delegate
- skill-tools.ts: skill, skill_mcp
- feature-tools.ts: task system (conditional), hashline (conditional)

coordinator keeps trimToolsToCap, LOW_PRIORITY_TOOL_ORDER, ToolRegistryResult.

Also: block subagent_type=sisyphus-junior with clearer error pointing to
category parameter which now routes to openfang agents.
