+++ src/plugin/tool-registry.ts
@@ -1,42 +1,20 @@
 import type { ToolDefinition } from "@opencode-ai/plugin"
-import type { SkillLoadOptions } from "../tools/skill/types"
-
-import type {
-  AvailableCategory,
-} from "../agents/dynamic-agent-prompt-builder"
+import type { AvailableCategory } from "../agents/dynamic-agent-prompt-builder"
 import type { OhMyOpenCodeConfig } from "../config"
 import { isInteractiveBashEnabled } from "../create-runtime-tmux-config"
 import type { PluginContext, ToolsRecord } from "./types"
-
-import {
-  builtinTools,
-  createBackgroundTools,
-  createCallOmoAgent,
-  createLookAt,
-  createSkillMcpTool,
-  createSkillTool,
-  createGrepTools,
-  createGlobTools,
-  createAstGrepTools,
-  createSessionManagerTools,
-  createDelegateTask,
-  discoverCommandsSync,
-  interactive_bash,
-  createTaskCreateTool,
-  createTaskGetTool,
-  createTaskList,
-  createTaskUpdateTool,
-  createHashlineEditTool,
-  createA2aDelegateTool,
-  DEFAULT_OPENFANG_BASE_URL,
-} from "../tools"
-import { getMainSessionID } from "../features/claude-code-session-state"
 import { filterDisabledTools } from "../shared/disabled-tools"
 import { isTaskSystemEnabled, log } from "../shared"
-
 import type { Managers } from "../create-managers"
 import type { SkillContext } from "./skill-context"
 import { normalizeToolArgSchemas } from "./normalize-tool-arg-schemas"
+import {
+  createSearchAndSessionTools,
+  createBackgroundAndBashTools,
+  createDelegationAndAgentTools,
+  createSkillAndMcpTools,
+  createFeatureTools,
+} from "./tools"
 
 export type ToolRegistryResult = {
   filteredTools: ToolsRecord
@@ -116,111 +94,13 @@ export function createToolRegistry(args: {
     availableCategories,
     interactiveBashEnabled = isInteractiveBashEnabled(),
   } = args
-  const backgroundTools = createBackgroundTools(managers.backgroundManager, ctx.client)
-  const callOmoAgent = createCallOmoAgent(
-    ctx,
-    managers.backgroundManager,
-    pluginConfig.disabled_agents ?? [],
-    pluginConfig.agents,
-    pluginConfig.categories,
-  )
-
-  const isMultimodalLookerEnabled = !(pluginConfig.disabled_agents ?? []).some(
-    (agent) => agent.toLowerCase() === "multimodal-looker",
-  )
-  const lookAt = isMultimodalLookerEnabled ? createLookAt(ctx) : null
-
-  const delegateTask = createDelegateTask({
-    manager: managers.backgroundManager,
-    client: ctx.client,
-    directory: ctx.directory,
-    userCategories: pluginConfig.categories,
-    agentOverrides: pluginConfig.agents,
-    gitMasterConfig: pluginConfig.git_master,
-    sisyphusJuniorModel: pluginConfig.agents?.["sisyphus-junior"]?.model,
-    browserProvider: skillContext.browserProvider,
-    disabledSkills: skillContext.disabledSkills,
-    availableCategories,
-    availableSkills: skillContext.availableSkills,
-    sisyphusAgentConfig: pluginConfig.sisyphus_agent,
-    syncPollTimeoutMs: pluginConfig.background_task?.syncPollTimeoutMs,
-    onSyncSessionCreated: async (event) => {
-      log("[index] onSyncSessionCreated callback", {
-        sessionID: event.sessionID,
-        parentID: event.parentID,
-        title: event.title,
-      })
-      await managers.tmuxSessionManager.onSessionCreated({
-        type: "session.created",
-        properties: {
-          info: {
-            id: event.sessionID,
-            parentID: event.parentID,
-            title: event.title,
-          },
-        },
-      })
-    },
-  })
-
-  const getSessionIDForMcp = (): string | undefined => getMainSessionID()
-
-  const skillMcpTool = createSkillMcpTool({
-    manager: managers.skillMcpManager,
-    getLoadedSkills: () => skillContext.mergedSkills,
-    getSessionID: getSessionIDForMcp,
-  })
-
-  const commands = discoverCommandsSync(ctx.directory, {
-    pluginsEnabled: pluginConfig.claude_code?.plugins ?? true,
-    enabledPluginsOverride: pluginConfig.claude_code?.plugins_override,
-  })
-  const skillTool = createSkillTool({
-    commands,
-    skills: skillContext.mergedSkills,
-    mcpManager: managers.skillMcpManager,
-    getSessionID: getSessionIDForMcp,
-    gitMasterConfig: pluginConfig.git_master,
-    browserProvider: skillContext.browserProvider,
-    nativeSkills: "skills" in ctx ? (ctx as { skills: SkillLoadOptions["nativeSkills"] }).skills : undefined,
-  })
-
-  const taskSystemEnabled = isTaskSystemEnabled(pluginConfig)
-  const taskToolsRecord: Record<string, ToolDefinition> = taskSystemEnabled
-    ? {
-        task_create: createTaskCreateTool(pluginConfig, ctx),
-        task_get: createTaskGetTool(pluginConfig),
-        task_list: createTaskList(pluginConfig),
-        task_update: createTaskUpdateTool(pluginConfig, ctx),
-      }
-    : {}
-
-  const hashlineEnabled = pluginConfig.hashline_edit ?? false
-  const hashlineToolsRecord: Record<string, ToolDefinition> = hashlineEnabled
-    ? { edit: createHashlineEditTool(ctx) }
-    : {}
-
-  const a2aDelegateTools = createA2aDelegateTool(
-    ctx,
-    pluginConfig.openfang?.base_url ?? DEFAULT_OPENFANG_BASE_URL,
-  )
 
   const allTools: Record<string, ToolDefinition> = {
-    ...builtinTools,
-    ...createGrepTools(ctx),
-    ...createGlobTools(ctx),
-    ...createAstGrepTools(ctx),
-    ...createSessionManagerTools(ctx),
-    ...backgroundTools,
-    call_omo_agent: callOmoAgent,
-    ...(lookAt ? { look_at: lookAt } : {}),
-    task: delegateTask,
-    skill_mcp: skillMcpTool,
-    skill: skillTool,
-    ...(interactiveBashEnabled ? { interactive_bash } : {}),
-    ...taskToolsRecord,
-    ...hashlineToolsRecord,
-    ...a2aDelegateTools,
+    ...createSearchAndSessionTools(ctx),
+    ...createBackgroundAndBashTools(managers.backgroundManager, ctx.client, interactiveBashEnabled),
+    ...createDelegationAndAgentTools({ ctx, pluginConfig, managers, skillContext, availableCategories }),
+    ...createSkillAndMcpTools({ ctx, pluginConfig, managers, skillContext }),
+    ...createFeatureTools(ctx, pluginConfig),
   }
 
   for (const toolDefinition of Object.values(allTools)) {
@@ -234,8 +114,7 @@ export function createToolRegistry(args: {
     trimToolsToCap(filteredTools, maxTools)
   }
 
-  return {
-    filteredTools,
-    taskSystemEnabled,
-  }
+  const taskSystemEnabled = isTaskSystemEnabled(pluginConfig)
+
+  return { filteredTools, taskSystemEnabled }
 }
